requirements.txt
Django==3.2.13
django-phonenumber-field==6.1.0
djangorestframework==3.13.1

gunicorn==20.1.0

phonenumbers==8.12.49
psycopg2-binary==2.9.1
pycountry==22.3.5

redis==3.0.0
